rename files, update docs, and update config for schema-builder merge
