+++ .gitignore
@@ -4,6 +4,7 @@
 /node_modules/
 /.tx/tmp/
 npm-debug.log
+.coverage
 
 transifex.auth
 interim
