+++ .vscode/settings.json
@@ -4,37 +4,37 @@
             "fileMatch": [
                 "data/fields/**/*.json"
             ],
-            "url": "./node_modules/@ideditor/schema-builder/schemas/field.json"
+            "url": "./schemas/field.json"
         },
         {
             "fileMatch": [
                 "data/presets/**/*.json"
             ],
-            "url": "./node_modules/@ideditor/schema-builder/schemas/preset.json"
+            "url": "./schemas/preset.json"
         },
         {
             "fileMatch": [
                 "data/preset_categories/**/*.json"
             ],
-            "url": "./node_modules/@ideditor/schema-builder/schemas/preset_category.json"
+            "url": "./schemas/preset_category.json"
         },
         {
             "fileMatch": [
                 "data/deprecated.json"
             ],
-            "url": "./node_modules/@ideditor/schema-builder/schemas/deprecated.json"
+            "url": "./schemas/deprecated.json"
         },
         {
             "fileMatch": [
                 "data/discarded.json"
             ],
-            "url": "./node_modules/@ideditor/schema-builder/schemas/discarded.json"
+            "url": "./schemas/discarded.json"
         },
         {
             "fileMatch": [
                 "data/preset_defaults.json"
             ],
-            "url": "./node_modules/@ideditor/schema-builder/schemas/preset_defaults.json"
+            "url": "./schemas/preset_defaults.json"
         }
     ],
     "files.exclude": {
