+++ CONTRIBUTING.md
@@ -42,13 +42,13 @@ You are highly welcome to help this project by submitting pull requests!
 
 ### Overview and General Structure
 
-Detailed documentation for the data format used in this repository is located with the [schema-builder](https://github.com/ideditor/schema-builder) package, which is the technical basis of this project.
+Detailed documentation for the data format used in this repository is located in [./SCHEMA.md](./SCHEMA.md).
 
 To make a change, update the corresponding file within the `data` folder: The `presets` contain a representation of OpenStreetMap's [map features](https://wiki.openstreetmap.org/wiki/Map_Features), and the `fields` are their properties. In addition, the tagging schema contains a few `categories` of presets and a list of `deprecated` and `discardable` tags.
 
 ### Icons
 
-Icons from different sources (_icon sets_) can be used in the tagging schema. Head over to the [dedicated page](https://github.com/ideditor/schema-builder/blob/main/ICONS.md#icons) about how to use them.
+Icons from different sources (_icon sets_) can be used in the tagging schema. Head over to the [ICONS.md](./ICONS.md) about how to use them.
 
 ### Info-`i`
 
@@ -57,7 +57,7 @@ Icons from different sources (_icon sets_) can be used in the tagging schema. He
 iD and other tools provide users with a way to learn more about the main tag of a preset. It is important to provide good information in this information panel. Here are a few notes on how to do this:
 - Does your tag have a OSM Wiki data item? Click the small pencil icon next to the text to open the data item on the OSM Wiki. Improve this wording if needed. If the data item is missing, [learn more about how to add it in "Current methods for creating new items"](https://wiki.openstreetmap.org/wiki/Data_items#Item_creation_process).
 - Does your tag have a Wiki page with a good image?
-- Your preset might need [a `reference` property](https://github.com/ideditor/schema-builder?tab=readme-ov-file#reference) to force the system to use a specific tag for the information section.
+- Your preset might need [a `reference` property](./SCHEMA.md#reference) to force the system to use a specific tag for the information section.
 
 ### Integration Testing With iD
