+++ SCHEMA.md
@@ -1,156 +1,6 @@
-[![build](https://github.com/ideditor/schema-builder/workflows/build/badge.svg)](https://github.com/ideditor/schema-builder/actions?query=workflow%3A%22build%22)
-[![npm version](https://badge.fury.io/js/%40ideditor%2Fschema-builder.svg)](https://badge.fury.io/js/%40ideditor%2Fschema-builder)
-
-# schema-builder
-
-This package lets you define and compile OpenStreetMap presets, fields, and other tagging
-info into the format expected by the [iD editor](https://github.com/openstreetmap/iD).
-
-The [iD Tagging Schema](https://github.com/openstreetmap/id-tagging-schema) project uses
-this to manage iD's tags. You can use it to create a custom schema for your own iD instance.
+# Schema
 
 ## Usage
-
-### Building Distribution Data
-
-To validate your source data and compile output files for iD (i.e. when releasing a new schema version):
-
-```JS
-const schemaBuilder = require('@ideditor/schema-builder');
-schemaBuilder.buildDist({
-  inDirectory: 'data',
-  interimDirectory: 'interim',
-  outDirectory: 'dist',
-  sourceLocale: 'en',
-  taginfoProjectInfo: {
-    name: 'IntrepiD',
-    description: 'iD editor, but adventurous.',
-    project_url: 'https://example.com/IntrepiD',
-    contact_name: 'J. Maintainer',
-    contact_email: 'maintainer@example.com'
-  }
-});
-```
-
-The following options are optional:
-
-- `inDirectory`: `string`, The relative directory of the source data files. Defaults to `data`.
-- `interimDirectory`: `string`, The relative directory of files needed during development but not for distribution. Be aware that
-everything in this directory will be overwritten when building. Defaults to `interim`.
-- `outDirectory`: `string`, The relative directory of the built data files intended for distribution. Be aware that
-everything in this directory will be overwritten when building. Defaults to `dist`.
-- `sourceLocale`: `string`, The code of the language/locale used for the translatable strings in the data files. Defaults to `en`.
-- `taginfoProjectInfo`: `object`, Project metadata required by TagInfo ([Wiki](https://wiki.openstreetmap.org/wiki/Taginfo/Projects)). If this info is not provided, the `taginfo.json` file will not be built. See the [schema](https://github.com/taginfo/taginfo-projects/blob/master/taginfo-project-schema.json) for more details. The generated taginfo.json will use the following mnemonics to give context to the generated [description on taginfo](https://taginfo.openstreetmap.org/projects/id_editor#tags):
-  - 🄿: [preset](https://github.com/openstreetmap/id-tagging-schema/tree/main/data/presets)
-  - 🄵: [field](https://github.com/openstreetmap/id-tagging-schema/tree/main/data/fields)
-  - 🄵🅅: field value
-  - 🄳: [deprecated tag](https://github.com/openstreetmap/id-tagging-schema/blob/main/data/deprecated.json)
-  - 🄳🄳: [discarded tag](https://github.com/openstreetmap/id-tagging-schema/blob/main/data/discarded.json)
-- `processPresets`: `function(presets)`, An opportunity to edit the built presets.
-- `processFields`: `function(fields)`, An opportunity to edit the built fields.
-- `processCategories`: `function(categories)`, An opportunity to edit the built preset categories.
-- `listReusedIcons`: `boolean` or `number`, If true, icons used by multiple searchable presets will be listed. If a number, icons used more than that number of times are listed. Defaults to `false`.
-
-You can also include options from `schemaBuilder.fetchTranslations()` in order to
-download translation files at the same time as compiling data.
-
-### Building Development Data
-
-To validate your source data and compile files needed during development:
-
-```JS
-const schemaBuilder = require('@ideditor/schema-builder');
-schemaBuilder.buildDev({
-  inDirectory: 'data',
-  interimDirectory: 'interim',
-  sourceLocale: 'en'
-});
-```
-
-The following options are identical to those for `schemaBuilder.buildDist()`:
-- `inDirectory`
-- `interimDirectory`
-- `sourceLocale`
-- `processPresets`
-- `processFields`
-- `processCategories`
-- `listReusedIcons`
-
-### Validating Data
-
-To validate your source data without compiling anything:
-
-```JS
-const schemaBuilder = require('@ideditor/schema-builder');
-schemaBuilder.validate({
-  inDirectory: 'data'
-});
-```
-
-The following options are identical to those for `schemaBuilder.buildDist()`:
-- `inDirectory`
-- `processPresets`
-- `processFields`
-- `processCategories`
-- `listReusedIcons`
-
-### Fetching Translations
-
-To download locale files from Transfiex:
-
-```JS
-const schemaBuilder = require('@ideditor/schema-builder');
-schemaBuilder.fetchTranslations({
-  outDirectory: 'dist',
-  sourceLocale: 'en',
-  translOrgId: 'openstreetmap',
-  translProjectId: 'intrepid',
-  translResourceIds: ['presets'],
-  translReviewedOnly: ['de', 'es']
-});
-```
-
-The following options are required:
-
-- `translOrgId`: `string`, The ID of the Transfiex organization where the translation project is hosted.
-- `translProjectId`: `string`, The ID of the Transfiex project within the organization where the schema resource is translated.
-
-The following options are optional:
-
-- `translResourceIds`: `[string]`, The IDs of the resources to download. Defaults to `['presets']`.
-- `translCredentials`: `{ user: string, password: string }`, Your Transifex API credentials.
-Defaults to those stored as JSON in a `transifex.auth` file in your working directory.
-- `translReviewedOnly`: `boolean` or `[string]`, If `true`, only reviewed translations are included.
-If `false`, all translations are included. If an array of locale codes, only reviewed
-translations are included for those specified locale codes, while all translations are included
-for the remaining locales.
-- `outDirectory`: `string`, Same as the `outDirectory` option for `schemaBuilder.buildDist()`.
-- `sourceLocale`: `string`, Same as the `sourceLocale` option for `schemaBuilder.buildDist()`.
-
-## Source Files
-
-Your `inDirectory` folder (`data` by default) should contain your source files with this structure:
-
-```
-data/
-    categories/
-        category1.json
-        category2.json
-        ...
-    fields/
-        field1.json
-        field2.json
-        ...
-    presets/
-        preset1.json
-        preset2.json
-        ...
-    defaults.json
-    deprecated.json
-    discarded.json
-```
-
-The format for each file is defined in the [`schemas`](schemas) directory.
 
 ### Presets
