+++ package.json
@@ -13,15 +13,31 @@
   "main": "build.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "lint": "prettier --check data",
-    "lint:fix": "prettier --write data",
+    "lint": "eslint && prettier --check data",
+    "lint:fix": "eslint --fix && prettier --write data",
     "build": "node scripts/build.js",
     "dist": "node scripts/dist.js",
-    "test": "node scripts/test.js"
+    "test": "node scripts/test.js && vitest run"
   },
   "devDependencies": {
-    "@ideditor/schema-builder": "^7.0.0",
+    "@eslint/js": "^10.0.1",
+    "@transifex/api": "^8.0.1",
+    "@types/taginfo-projects": "github:taginfo/taginfo-projects#6daafd2e30a86a8bfb9615855af2c2d73bd98279",
+    "@vitest/coverage-v8": "^4.1.5",
+    "@vitest/eslint-plugin": "^1.6.16",
+    "eslint": "^10.0.0",
+    "globals": "^17.3.0",
+    "js-yaml": "^5.0.0",
+    "json-schema-to-typescript-lite": "^15.0.0",
+    "jsonschema": "^1.1.0",
+    "marky": "^1.2.4",
+    "node-fetch": "^3.2.10",
     "prettier": "^3.2.5",
-    "unzipper": "^0.12.3"
+    "shelljs": "^0.10.0",
+    "unzipper": "^0.12.3",
+    "vitest": "^4.1.5"
+  },
+  "engines": {
+    "node": ">=22"
   }
 }
