+++ schemas/field.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "https://github.com/ideditor/schema-builder/raw/main/schemas/field.json",
+    "$id": "https://github.com/openstreetmap/id-tagging-schema/raw/main/schemas/field.json",
     "title": "Field",
     "description": "A reusable form element for presets",
     "type": "object",
