+++ schemas/preset.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "https://github.com/ideditor/schema-builder/raw/main/schemas/preset.json",
+    "$id": "https://github.com/openstreetmap/id-tagging-schema/raw/main/schemas/preset.json",
     "title": "Preset",
     "description": "Associates an icon, form fields, and other UI with a set of OSM tags",
     "type": "object",
