+++ schemas/preset_defaults.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "https://github.com/ideditor/schema-builder/raw/main/schemas/preset_defaults.json",
+    "$id": "https://github.com/openstreetmap/id-tagging-schema/raw/main/schemas/preset_defaults.json",
     "title": "Default presets",
     "description": "Directory of the IDs of the presets or categories to be shown in the default list for each geometry type.",
     "type": "object",
