+++ scripts/lib/index.js
