+++ scripts/lib/references.js
