+++ scripts/lib/translations.js
