+++ vitest.config.js
@@ -2,11 +2,10 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    globals: true,
     coverage: {
       enabled: true,
       provider: 'v8',
-      include: ['lib/*.js'],
+      include: ['script/*.js'],
       reportsDirectory: '.coverage',
     },
   },
